pyproject.toml
[project]
name = "zed-base"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask (>=3.1.3,<4.0.0)",
    "pyqt5 (>=5.15.11,<6.0.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "flask-login (>=0.6.3,<0.7.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-socketio (>=5.3.6,<6.0.0)",
    "flask-cors (>=4.0.0,<5.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "authlib (>=1.6.9,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "python-socketio (>=5.11.0,<6.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "psycopg[binary] (>=3.3.3,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
